Post-hoc Within Group Comparisons of Differences in a Survival Outcome

Performs pairwise comparisons of the levels of one categorical variable within each level of another categorical variable adjusting for multiplicity.

Usage

survcomp(dat, patid, x_c, x_n, dt_start, dt_outcome, dt_end, unit = "month")

Arguments

dat

a data frame.

patid

an integer vector of the subject IDs.

x_c

the name(s) of categorical variable(s).

x_n

the name(s) of numeric variable(s).

dt_start

the name of the date of the starting point.

dt_outcome

the name of the date of the survival outcome.

dt_end

the name of the date of the endpoint.

unit

the unit of measurement for the time variable; options are: "day", "month", "year".

Value

a summary_emm object and a data frame of pairwise comparisons, confidence intervals, test statistics, and p-values.

Examples

# Read in the sample data
path <- system.file("extdata", "bcsurg.csv", package = "survcomp2")
dat <- read.csv(path)
emm_contrasts <- survcomp(
dat = dat, patid = "ID..",
x_c = c("Groups", "Breast_surgery_code"),
x_n = "AGE_agedx", dt_start = "DT_dxdate2",
dt_outcome = "DT_dod",
dt_end = "DT_date_last_seen"
)
# Only keep the statistically significant pairwise comparisons
emm_contrasts[emm_contrasts$p.value < 0.05, ]
#>  contrast          Breast_surgery_code ratio     SE  df asymp.LCL asymp.UCL
#>  Groups0 / Groups2 0                   0.435 0.0918 Inf     0.245     0.774
#>  Groups0 / Groups2 1                   0.430 0.0945 Inf     0.236     0.783
#>  Groups2 / Groups4 1                   4.364 2.2824 Inf     1.048    18.175
#>  null z.ratio p.value
#>     1  -3.944  0.0008
#>     1  -3.838  0.0012
#>     1   2.817  0.0390
#> 
#> Confidence level used: 0.95 
#> Conf-level adjustment: tukey method for comparing a family of 5 estimates 
#> Intervals are back-transformed from the log scale 
#> P value adjustment: tukey method for comparing a family of 5 estimates 
#> Tests are performed on the log scale